backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
numpy>=1.26.0
scipy>=1.12.0
openai>=1.10.0
google-generativeai>=0.3.0
httpx>=0.26.0
python-multipart>=0.0.6
reportlab>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=7.4.0
pytest-asyncio>=0.23.0
